requirements.txt
fastapi
uvicorn
yfinance
supabase>=2.3.0
pytest>=8.0.0
pytest-asyncio>=0.23.0
httpx>=0.26.0
pytest-cov>=4.1.0
python-dotenv>=1.0.0
paypalrestsdk>=1.13.3
requests>=2.31.0
pandas
cachetools
python-multipart
feedparser
requests